Added some Docker support
Added a Dockerfile to create a Docker image. You will need to run:

docker build -t  samply/transfyr . --no-cache

Additionally, a docker-compose.yml file was added to allow users to
try transfyr out. This is basically for demonstration purposes.

Finally, application.yml was modified so that some of the configurations
can be overwritten by environment variables.