Convert from BBMRI FHIR to Beacon
This initial commit is for Patients/individuals only, it does not cover
samples.

Also, Observation information is not copied over to individuals, that
needs to be implemented in a future iteration.

New code was added to BatchConfiguration.java and resources/application.yml
in order to realize the new converter.

Additionally, a new processor and a new writer needed to be written:

processor/BbmriBundleToBeaconProcessor.java
writer/BeaconIndividualWriter.java

The following were copied with little or no change from the old
Transfair implementation, meaning that they already passed an
earlier code review:

mapper/bbmri2beacon/*
models/beacon/*
util/JsonSerializer.java
writer/BeaconFileSaver.java
writer/BeaconMongoSaver.java

The one major change in the copied files was in:

mapper/bbmri2beacon/Bbmri2BeaconIndividual.java

which had to be adapted to take a Bundle as input, rather than directly
connecting to the FHIR store.