Added conversion of samples.
Some changes were needed to class names to distinguish between classes
dealing with individuals and classes dealing with biosamples.

I added an extra step to the bbmri2beacon profile in BatchConfiguration
to generate BFF for biosamples.

Additionally, I added processor and writer classes for biosamples.

Bbmri2BeaconBiosample is the class where the actual conversion work
is done.

All of the new infrastructure works in a very similar way to the
existing infrastructure for individuals and the same coding conventions
have been used.