Added import from antimicrobial resistance data
The EHDS2 project has initiated a pilot project to build a search
platform for antimicrobial resistance data.

The updates to TransFAIR in this commit are intended to allow the import of data
from a CSV file containing the relevant data, one observation per
row.

Most of the new code is in FhirAmtReader.java, but changes to
BatchConfiguration.java were also needed to make this work.

A new dependency has been added to pom.xml to get the Apache CSV
parser.

The Dockerfile had to be modified to nail down the version of the
Maven image, because I was getting errors due to version conflicts.