Removed unused Spring Boot tag
TransFAIR was crashing occasionally, complaining of a
ConflictingBeanDefinitionException

I think this might be because the application contains two beans with
the same name.

Anyway, I discovered that FhirBundleBuilder had a Spring Component
tag, but this wasn't actually being used. So I commented it out and
recompiled and then tested it.

Sure enough, the problem went away.