Stop the generation of identical Observations
It turns out that the test data contains several hundred identical
observations, which have come to light since implementing an OD based
on an attribute hash.

That is a problem for Blaze, which throws an error if you try to insert
a copied Observation ID.

So in this commit, we simply skip repeats of an Observation.